Restore signal generation logic in signals.py
Restored import statements and function logic for signal generation.

+++ signals.py
@@ -87,7 +87,7 @@ def get_signals(symbol):
         "MACD": float(latest["MACD"]),
         "MACD_SIGNAL": float(latest["MACD_SIGNAL"]),
 
-        "SUPERTREND": float(latest["ST_DIRECTION"]),
+        "SUPERTREND": int(latest["ST_DIRECTION"]),
 
         "Volume": float(latest["Volume"]),
         "AVG_VOLUME": float(latest["AVG_VOLUME"]),
